run script update && generalization

+++ scripts/run.sh
@@ -1,15 +1,25 @@
 #!/bin/bash -x
 
-for i in block fs mm security virt crypto net sound
+kver="$1"
+
+kfile="linux-${kver}.tar.bz2"
+
+rver="$2"
+
+if [[ ! -r "$kfile" ]]
+then
+   exit 1
+fi
+
+for i in arch block fs mm security virt crypto net sound
 do
-   LDV_DEBUG=100 ldv-manager "envs=linux-3.2.22.tar.bz2" kernel_driver=1 "drivers=$i/" "rule_models=113_1a" | tee log-$i.txt
+   LDV_DEBUG=100 ldv-manager "envs=${kfile}" kernel_driver=1 "drivers=${i}/" "rule_models=${rver}" | tee "log-${i}.txt"
    rm -fr ./work/
 done
 
-for i in  $(find /home/work/workspace/linux-stable/drivers/ -maxdepth 1 -type d | cut -d '/' -f 6- | tail -n +2 | less)
+for i in $(find ./inst/current/envs/linux-${kver}/linux-${kver}/ -maxdepth 1 -type d | cut -d '/' -f 6- | tail -n +2 | less)
 do
-   LDV_DEBUG=100 ldv-manager "envs=linux-3.2.22.tar.bz2" kernel_driver=1 "drivers=$i/" "rule_models=113_1a" | tee log-$i.txt
+   LDV_DEBUG=100 ldv-manager "envs=${kfile}" kernel_driver=1 "drivers=${i}/" "rule_models=${rver}" | tee "log-${i}.txt"
    rm -fr ./work/
 done
-
